+++ CHANGELOG.md
@@ -10,7 +10,9 @@ Versioning].
 ## [Unreleased] <!-- #release:date -->
 
 * Add support for Plan metadata.
-* Bump MSRV to 1.67.
+* Add support for creating customer ledger entries.
+* Add support for enumerating customer credit balances.
+* Bump MSRV to 1.70.
 
 ## [0.6.0] - 2023-05-12